requirements.txt
fastapi>=0.110
uvicorn>=0.30
pydantic>=2.6
openai>=1.0
python-dotenv>=1.0
python-multipart>=0.0.9
requests>=2.31
httpx>=0.27
